use decomposition to make the code more understandable
